// frontend/src/app/api/model/matchPaymentDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MatchPaymentDto { 
    tr?: number;
    userMatricule?: string;
    matchId?: number;
    amount?: number;
    paymentDate?: string;
    status?: string;
    paymentMethod?: string;
}